refactor(docker): remove unused models and utilities for cleaner codebase
